Refactor A1 agent to A2 and SubA1 to A1
